Fix issues in TODO, disable logging
The input in `input/overlap` is now clustered correctly. Fewer events
are created and processed, by
 (1) not creating OUT_OF_CELL events initially for when glyphs grow out
     of the root QuadTree cell; and
 (2) not creating OUT_OF_CELL events in some cases, when handling
     OUT_OF_CELL events. This was already done in other places.
The nested merging loop now continues even when the next event does not
involve the merged glyph. This fixes the problem in `input/overlap`. It
makes sense, as sometimes stale events need to be handled before events
of interest pop up. The check is therefore delayed.

This commit also undoes the inversion of the Y axis. Many places in the
code (sides of cells, et cetera) assume an inverted Y axis. I deem it
too much trouble for little reward to invert everything.