Add glyphs to cells less eagerly
In the scenario in the newly created `input/careful-out`, a glyph grows
into a neighboring cell, but not in all neighbors along the side of the
cell it grows from. This is now handled properly.

The algorithm now also produces some statistics about number of
processed events through the algorithm. Adjusted log level to see that.