Fix overlap in `input/overlap-2`
There was an issue with `QuadTree#getNeighbors`, caused by a fault in
`QuadTree#getLeaves`. Or, two faults, to be precise. One was that leaves
from the wrong side of the neighboring cell were retrieved. The other,
in `getLeaves`, was that a cell just barely in range (overlap in 1
point) would be included, while that was not intended.