Restore previous behavior behind a flag
Adding only the first merge is theoretically dangerous and could result
in overlap in the clustering. The previous behavior of adding a merge
event for all glyphs in the same QuadTree cells is now restored, and can
be enabled through a `static final` flag in `AgglomerativeClustering`.

I investigated an alternative approach, wherein glyphs track which
glyphs think they merge with them first. Upon merging with a glyph,
other glyphs that are still tracking it then update their merge events.
This appeared to perform suboptimally.