Reintroduce tracking solution
As mentioned in previous commit message, I had investigated an
alternative solution that appeared to perform very badly. This commit
reintroduces that approach, but executed in a slightly better way and
behind a flag. It turns out that this approach, like `ROBUST` merge
event handling, needs a much lower maximum number of glyphs per cell to
make it work. It does outperform the `ROBUST` approach, greatly.