Recursively join QuadTree cells
Instead of joining exactly once, test if after joining, another join can
be performed recursively. This is usually not possible, but when the
maximum number of glyphs per cell is smaller, this will occur more
often. It also seems to happen more often in larger QuadTrees in
general. Effects on running time are negligible usually, and minor for
large QuadTrees, but noticeable.

Closes issue #1.