Create fewer (unnecessary) merge events
When merging a glyph, it is better to wait until the full merge has
completed before creating merge events and updating the merge events of
glyphs that were tracking a merged glyphs. The reason is that nested
merges may cause more updates to be needed, and duplicate merge events
to be created. This is exactly what this commit implements.

Closes issue #4.