Start work on finding good constants
Problem I ran into is that clustering the same data set again will
result in very strange errors in the resulting clustering, and sometimes
in errors about the QuadTree splitting too often. This is a WIP commit,
I am still tracking down the issue(s). For now, it is easy to reopen a
file via the GUI, and reopening appears to be working correctly. It's
just the combination of clustering, reopening and clustering that gives
bad results. The investigation continues tomorrow.