Make merge recorder stream sequential
After noticing that the number of glyphs that is recorded each time is
at most 12 or so, for the current settings (10 glyphs per cell), I
realized the overhead of parallellizing the stream is not worth it. This
turns out to be accurate; clustering is about 1 second faster now. The
Trove data set can be clustered in about 0.65 seconds.