Refactor grow functions; split shape and speed
A lot of code was duplicated in the various grow function
implementations. Splitting the part of a grow function that determines
its shape from the part that determines the speed at which glyphs grow
eliminates this code duplication.

I initially attempted different ways of doing the split, that led grow
functions to have generics parameters determining their shape and speed
functions. Eventually I decided not to pursue that route because the
shape and speed parts need certain parts of each other. It became
impractical, if not impossible, to get this right. The current setup has
the shape and speed parts of grow functions refer to their parent grow
function, which enables them to use each other.