Fix half a bug involving missing works
There are now too many works in the resulting clustering. Problems seem
to happen in the phase before actual clustering starts (resolving
overlapping glyphs with collision time < 0). A full-blown check is built
in now, but disabled with a constant because of its impact on running
time of the clustering.

Minor improvements to the drawing of glyphs in the debugging GUI are
also in this commit (smaller labels, but split over lines; this avoids
overlap in more cases and thus makes labels better legible).

Finally, the view onto hierarchical clusterings now only includes merges
that happened after t = 0, to make the view more consistent. It should
not be needed to consider merges before t = 0 anyway.