Add new glyph generator, fix out of cell bug
It is now possible to generate data sets where there is a single big
glyph in the center, with a ring of uniformly distributed glyphs around
it. This could be a bad case for our clustering algorithm.

There was a bug where the big glyph would overlap glyphs on the ring,
caused by not enough out of cell events being popped into the queue.
This is fixed by adding one `popOutOfCellInto` call in the
`QuadTreeClusterer`.

I discovered and resolved the above bug thanks to (re)new(ed) tracking
functionality. Glyphs can be tracked by clicking them in the GUI prior
to clustering. Events involving tracked glyphs will be logged about
extensively, even if logging is otherwise disabled.