Protect against out of memory in batch mode
We don't do naive on large data sets, it simply won't fly. A nicer
approach would be to track used memory and kill the thread if it runs
out, or even catch the OutOfMemoryError and handle it ... But that
doesn't work and cannot be done in Java. Definitely not quickly.